Version 2

+++ simpleplot/ploting/plot_data_types/distribution_data.py
@@ -28,7 +28,7 @@ from ..plot_geometries.surfaces import QuadSurface
 from ..plot_geometries.points   import Point
 from ...pyqtgraph.pyqtgraph     import functions
 from ...pyqtgraph.pyqtgraph.opengl.MeshData import MeshData
-from ...model.node   import SessionNode
+from ...models.session_node   import SessionNode
 
 class DistributionData(PlotData, SessionNode): 
     '''
@@ -58,7 +58,7 @@ class DistributionData(PlotData, SessionNode):
         for i,value in enumerate(self._axes):
             if value in kwargs.keys():
                 if isinstance(kwargs[value],np.ndarray) or isinstance(kwargs[value],list):
-                    elements[i] = np.array(kwargs[value])
+                    elements[i] = np.array(kwargs[value]).astype(np.float)
                     changed[self._axes.index(value)] = True
 
         if elements[self._axes.index('data')] is None:
