+++ simpleplot/ploting/plot_data_types/volume_data.py
@@ -29,7 +29,7 @@ from ..plot_geometries.points   import Point
 from ..plot_geometries.shaders  import ShaderConstructor
 from ...pyqtgraph.pyqtgraph     import functions
 
-from ...model.node   import SessionNode
+from ...models.session_node   import SessionNode
 
 class VolumeData(PlotData, SessionNode): 
     '''
@@ -58,7 +58,7 @@ class VolumeData(PlotData, SessionNode):
         for i,value in enumerate(self._axes):
             if value in kwargs.keys():
                 if isinstance(kwargs[value],np.ndarray) or isinstance(kwargs[value],list):
-                    elements[i] = np.array(kwargs[value])
+                    elements[i] = np.array(kwargs[value]).astype(np.float)
                     changed[self._axes.index(value)] = True
 
         if elements[self._axes.index('data')] is None:
